Text Summarizer
Built a Text Summarizer which reduces a text file into short set of
words to convey the main meaning of the text file. Linguistic methods
were used to examine and interpret the file to generate a smaller file
by removing extraneous phrases so that it conveys the most important
information from the original text document.